Fix a potential migration error

+++ db/migrate/20140416142715_drop_institution_types_table.rb
@@ -1,5 +1,5 @@
 class DropInstitutionTypesTable < ActiveRecord::Migration
   def change
-    drop_table :institution_types
+    drop_table :institution_types if table_exists? :institution_types
   end
 end
